renaming qaPanel 1/2
