+++ src/components/qaPanel2/qaPanel2.css
